% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for a table}
\usage{
get_available_years(con, schema = "llds", table, year_col)
}
\arguments{
\item{con}{A DBI connection object.}

\item{schema}{Character. Schema name. Default: "llds".}

\item{table}{Character. Table name.}

\item{year_col}{Character. Name of the year column.}
}
\value{
Integer vector of available years, sorted ascending.
}
\description{
Queries a database table to find the distinct values of a year column.
Used to determine how to partition parallel downloads.
}
\examples{
\dontrun{
con <- create_connection()
years <- get_available_years(con, table = "LLDS_HDR_20240315HAC", year_col = "LANDYR")
safe_disconnect(con)
}
}
